% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_LGP.R
\name{ebnm_LGP_generator}
\alias{ebnm_LGP_generator}
\title{Generate a Custom `ebnm` Function for L-GP Smoothing}
\usage{
ebnm_LGP_generator(LGP_setup, link = "identity")
}
\arguments{
\item{LGP_setup}{A list of precomputed matrices and related quantities (e.g., from \code{LGP_setup}) required for L-GP smoothing.}

\item{link}{A character string specifying the link function to use. Default is \code{"identity"}.}
}
\value{
A function that performs L-GP smoothing with:
  - Posterior moments estimation.
  - Log-likelihood computation.
  - Posterior sampling.
}
\description{
Creates a custom `ebnm` function for fitting data using L-GP models.
}
\examples{
setup <- LGP_setup(t = seq(0, 1, length.out = 100))
ebgp <- ebnm_LGP_generator(setup)
result <- ebgp(x = rnorm(100), s = 0.1)

}
