% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_utility.R
\name{init_rank}
\alias{init_rank}
\title{Initialize Rank with Smoothness Estimation}
\usage{
init_rank(
  data,
  S = NULL,
  k_total = 5,
  ebnm_indep = NULL,
  var_type = NULL,
  verbose = 0,
  backfitting_steps = 0,
  convergence_fn = flashier::flash_conv_crit_max_chg,
  convergence_tol = 1e-04,
  smooth_term = "F",
  smooth_fun
)
}
\arguments{
\item{data}{A matrix of observed data.}

\item{S}{(Optional) A matrix of standard errors corresponding to the data in \code{data}. Defaults to \code{NULL}.}

\item{k_total}{An integer specifying the total number of factors to fit. Defaults to \code{5}.}

\item{ebnm_indep}{(Optional) A function for empirical Bayes estimation of independent factors. Defaults to \code{ebnm::ebnm_point_normal}.}

\item{var_type}{(Optional) An integer specifying the variance type. Defaults to \code{NULL}.}

\item{verbose}{An integer controlling verbosity. Set to \code{0} for no messages, or higher values for more detailed messages. Defaults to \code{0}.}

\item{backfitting_steps}{An integer specifying the number of backfitting iterations. If \code{0}, backfitting is skipped. Defaults to \code{0}.}

\item{convergence_fn}{A function defining the convergence criterion. Defaults to \code{flashier::flash_conv_crit_max_chg}.}

\item{convergence_tol}{A numeric value specifying the convergence tolerance for smooth factor fitting. Defaults to \code{1e-4}.}

\item{smooth_term}{A string indicating whether to estimate smoothness for factors (\code{"F"}) or loadings (\code{"L"}). Defaults to \code{"F"}.}

\item{smooth_fun}{A smoothing function to estimate the smoothness of the selected terms.}
}
\value{
A list containing:
\itemize{
  \item \code{fit}: A `flashier` fit object containing the fitted factors and loadings.
  \item \code{smoothness}: A numeric vector of estimated smoothness values for the selected terms (factors or loadings).
}
}
\description{
This function initializes the rank of factors or loadings in a dataset
using the `flashier` package and evaluates their smoothness. The function first fits
a specified number of independent factors and then applies a smoothing function to
estimate the smoothness of either the factors or loadings.
}
\details{
The function performs the following steps:
1. Fits a specified number of independent factors using \code{flashier::flash_greedy()}.
2. Optionally performs backfitting to refine the solution.
3. Applies a smoothing function to estimate the smoothness of the factors or loadings, as specified by \code{smooth_term}.
}
