% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Iyer_expression}
\alias{Iyer_expression}
\title{The (log-transformed) relative gene expression analyzed in Iyer et al. (1999).}
\format{
'Iyer_expression'
A data frame with 517 obs of 16 variables.
}
\usage{
Iyer_expression
}
\description{
The (log-transformed) relative gene expression data for 517 genes at 12 time points, measured
relative to the 0 minute time point.
}
\keyword{datasets}
