% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_LGP.R
\name{LGP}
\alias{LGP}
\title{Define the L-GP Object}
\usage{
LGP(scale)
}
\arguments{
\item{scale}{A numeric value representing the scale parameter of the L-GP.}
}
\value{
An object of class `"LGP"`, a data frame containing the scale parameter.
}
\description{
Creates an object from the one-parameter L-GP family.
}
\examples{
gp <- LGP(1.5)
print(gp)

}
