% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Matern_Spatial.R
\name{Matern}
\alias{Matern}
\title{Define the Matern GP Family}
\usage{
Matern(theta = NULL)
}
\arguments{
\item{theta}{A numeric value representing the log-transformed range parameter
(optional). Defaults to \code{NULL}.}
}
\value{
An object of class \code{"Matern"}, representing a Matern process with the
specified parameters.
}
\description{
Creates an object representing the Matern process. The Matern process
is parameterized by a log-transformed range parameter, \code{theta}, which defines
the spatial correlation scale.
}
\examples{
matern_gp <- Matern()
print(matern_gp)

}
