% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_utility.R
\name{plot_factors_and_loadings_1d}
\alias{plot_factors_and_loadings_1d}
\title{Plot Factors and Loadings from a Flashier Model fitted with LGP prior}
\usage{
plot_factors_and_loadings_1d(f_mod, t_vec, true_L = NULL, true_F = NULL)
}
\arguments{
\item{f_mod}{A flashier model object containing \code{L_pm} (loadings) and \code{F_pm} (factors).}

\item{t_vec}{A numeric vector representing the time points for the smooth loadings.}

\item{true_L}{A matrix of true loadings (optional). Defaults to \code{NULL}.}

\item{true_F}{A matrix of true factors (optional). Defaults to \code{NULL}.}
}
\value{
Generates plots showing the comparison between estimated and true loadings/factors for all components.
}
\description{
This function plots the normalized factors and loadings from a fitted
flashier model alongside the true normalized factors and loadings.
}
\examples{
set.seed(123)
# Simulated data
sim <- simulate_matrix(n = 200, p = 100, k_smooth = 2, k_ind = 0, sigma_E = 1)
f.mod <- flashier::flash_init(data = sim$Y) |>
  flashier::flash_greedy(Kmax = 2, ebnm_fn = ebnm::ebnm_point_normal)
plot_factors_and_loadings_1d(f_mod, sim$t_vec, sim$L_mat[, 1:2], sim$F_mat[, 1:2])

}
