% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Matern_Spatial.R
\name{ebnm_Matern_generator}
\alias{ebnm_Matern_generator}
\title{Generate a Custom `ebnm` Function for Matern Smoothing}
\usage{
ebnm_Matern_generator(
  locations,
  max.edge = NULL,
  alpha = 2,
  range_prob = 0.05,
  sigma_prob = 0.05,
  range_val = 2,
  sigma_val = 2
)
}
\arguments{
\item{locations}{A matrix of spatial locations, where each row represents a coordinate.}

\item{max.edge}{A numeric value defining the maximum edge length for the spatial mesh. Default is \code{NULL}.}

\item{alpha}{A numeric value specifying the smoothness parameter of the Matern GP. Default is \code{2}.}

\item{range_prob}{A numeric value specifying the prior probability for the range parameter. Default is \code{0.05}.}

\item{sigma_prob}{A numeric value specifying the prior probability for the standard deviation parameter. Default is \code{0.05}.}

\item{range_val}{A numeric value specifying the prior mean for the range parameter. Default is \code{2}.}

\item{sigma_val}{A numeric value specifying the prior mean for the standard deviation parameter. Default is \code{2}.}
}
\value{
A function that can be used to fit data using the Matern process The returned function includes:
\itemize{
  \item \code{posterior}: A data frame of posterior mean, variance, and second moments.
  \item \code{fitted_g}: The fitted Matern object.
  \item \code{log_likelihood}: The marginal log-likelihood of the model.
  \item \code{posterior_spatial_field}: Summary of the spatial random field.
  \item \code{mesh}: The spatial mesh used in the model.
  \item \code{inla_result}: The full result object from \code{INLA::inla()}.
}
}
\description{
Creates a custom `ebnm` function for fitting data using Matern process.
}
\examples{
# Example usage of the Matern GP model
locations <- matrix(runif(20), ncol = 2)  # Random 2D locations
matern_model <- ebnm_Matern_generator(locations = locations)

# Data and standard deviations
x <- rnorm(nrow(locations))
s <- rep(0.1, nrow(locations))

result <- matern_model(x = x, s = s)
str(result$posterior)

}
