% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Matern_Spatial.R
\name{Matern}
\alias{Matern}
\title{Define the Matern GP Family}
\usage{
Matern(theta = NULL)
}
\value{
An object of class `"Matern"`, representing a Matern process.
}
\description{
Creates an object representing the Matern process.
}
\examples{
matern_gp <- Matern()
print(matern_gp)

}
